♻️ refactor: 모임찾기 카드리스트 리팩토링
